Devide scripts into modules
